Scaled Satchel Speedloss feature
An optional change
usage:
sr_scaledSatchelRange 0 (off) or 1 (on)

if 0: classic, you step outside of the range and you get instantly
slowed to the minimum speed.
if 1: every second you lose % when you're outside of the range till you
reach the sr_scaledSatchelMinSpeed. The % is optional and you can
control it in sr_scaledSatchelSpeedLossPercent.